requirements.txt
aiohttp
websockets
lxml
discord.py>=0.16.0
